requirements.txt
streamlit==1.34.0
gtts==2.5.1
transformers==4.38.1
Pillow==10.0.1
torch==2.1.0
pyttsx3==2.90
